#Development-Mirek Car brands, style changed, and added security to car brands, and body when editing entities

+++ TypicalMirek_UsedCarDealer/Logic/Managers/BrandManager.cs
@@ -13,15 +13,18 @@ namespace TypicalMirek_UsedCarDealer.Logic.Managers
 {
     public class BrandManager : Manager, IBrandManager
     {
+        #region Properties
         private readonly IBrandRepository brandRepository;
+        private readonly ICarRepository carRepository;
+        #endregion
 
-        //private readonly UnitOfWork unitOfWork;
-
+        #region Constructors
         public BrandManager(IRepositoryFactory repositoryFactory)
         {
-           // unitOfWork = unitOfWorkFactory.Get<UnitOfWork>();
             brandRepository = repositoryFactory.Get<BrandRepository>();
+            carRepository = repositoryFactory.Get<CarRepository>();
         }
+        #endregion
 
         public IQueryable<Brand> GetAll()
         {
@@ -35,10 +38,11 @@ namespace TypicalMirek_UsedCarDealer.Logic.Managers
 
         public Brand Add(Brand brand)
         {
-            if (brandRepository.GetById(brand.Id) != null && brandRepository.GetAll().FirstOrDefault(b => b.Name.Equals(brand.Name)) != null)
+            if (brandRepository.GetById(brand.Id) != null || brandRepository.CheckIfBrandWithExactNameExists(brand.Name))
             {
                 return null;
             }
+
             brandRepository.Add(brand);
             brandRepository.Save();
             return brand;
@@ -47,7 +51,7 @@ namespace TypicalMirek_UsedCarDealer.Logic.Managers
         public Brand Modify(Brand brand)
         {
             var brandToModify = brandRepository.GetById(brand.Id);
-            if (brandToModify == null)
+            if (brandToModify == null || brandRepository.CheckIfBrandWithExactNameExists(brand.Name))
             {
                 return null;
             }
@@ -57,16 +61,24 @@ namespace TypicalMirek_UsedCarDealer.Logic.Managers
             return brand;
         }
 
-        public void Delete(Brand brand)
+        public bool Delete(Brand brand)
         {
-            brandRepository.Delete(brand);
-            brandRepository.Save();
+            if (brand != null)
+            {
+                if (!carRepository.CheckIfExistCarForBrandId(brand.Id))
+                {
+                    brandRepository.Delete(brand);
+                    brandRepository.Save();
+                    return true;
+                }
+            }
+            return false;
         }
 
         public void Dispose()
         {
             brandRepository.Dispose();
-            //unitOfWork.Dispose();
+            carRepository.Dispose();
         }
     }
 }
